scoring tests

+++ src/main/java/jevm/nutshell/engine/TextAnalyzer.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 public interface  TextAnalyzer {
 
-    public Map<String, Integer> getKeyWords(int n);
+    public Map<String, Double> getKeyWords(int n);
     public int scoreSentence(String sentence);
     public int scoreWord(String word);
